created list that takes userInput and tests it against a keyID # that has been assigned to each item (12) in total.

+++ Player.cs
@@ -7,7 +7,7 @@ namespace LemondStandTest
     
     public class Player
     {
-        public float wallet = 20f;
+        public double wallet = 20f;
         PlayerInventory Inventory = new PlayerInventory();
